backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Password Hashing
bcrypt==4.1.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# NLP & ML - Using compatible versions for Windows
spacy==3.7.2
scikit-learn==1.3.2

# LLM API
openai==1.6.1

# Web Scraping (Optional)
beautifulsoup4==4.12.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
Werkzeug==3.0.1

# Production Server
gunicorn==21.2.0
